app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
pgvector==0.2.4
sentence-transformers>=2.7.0
faiss-cpu==1.7.4
numpy==1.24.3
torch>=2.0.0
transformers>=4.30.0

